library/017-AI/近屿/aigc_learning_path/aigc_tm_level/A3/A3.3大模型RAG/requirements.txt
sparkdesk-api
websocket-client
openai>=1.0
markdown-it-py
faiss-cpu
langchain==0.3.13
pymupdf
numpy==1.24.3
matplotlib
python-dotenv
tiktoken==0.8.0
langchain-community==0.3.13
PyPDF2==3.0.1
jieba==0.42.1